[#10] rename: 폴더명 수정
